test(sync-server): add integration test infrastructure
